renamed user slice returned object
